// apps/core-service/src/app/modules/authentication/dto/SsoConfig.dto.ts
//=============================================================================
//Licensed Materials - Property of Bambee
//(C) Copyright Bambee 2022
//All Rights Reserved
//=============================================================================

/* istanbul ignore file */
export interface SsoConfigDTO {
  clientId: string;
  clientSecret: string;
  publicKey: string;
  privateKey: string;
}